Introduce `CodegenMember`
Also:
 * Make `FileRenderer` a class extending `CodegenMember`.
